Moved TargetPlatformHelper to ui package- (Task #25)
---
Task #25: Resolve target platform from tmodel editor
